+++ src/main/java/com/onlineshopthymeleaf/repository/CategoryRepository.java
@@ -4,5 +4,5 @@ import com.onlineshopthymeleaf.model.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 // CategoryRepository.java
-public interface CategoryRepository extends JpaRepository<Category, Long> {}
+public interface CategoryRepository extends JpaRepository<Category, Byte> {}
