+++ src/main/java/com/onlineshopthymeleaf/repository/ColorRepository.java
@@ -3,5 +3,5 @@ package com.onlineshopthymeleaf.repository;
 import com.onlineshopthymeleaf.model.Color;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ColorRepository extends JpaRepository<Color,Long> {
+public interface ColorRepository extends JpaRepository<Color,Byte> {
 }
