+++ src/main/java/com/onlineshopthymeleaf/repository/ProductRepository.java
@@ -9,11 +9,11 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface ProductRepository extends JpaRepository<Product, Long> {
+public interface ProductRepository extends JpaRepository<Product, Byte> {
     @Query("SELECT p FROM Product p LEFT JOIN FETCH p.images")
     List<Product> findAllWithImages();
     @Query("SELECT pi.images FROM Product pi WHERE pi.id = :productId")
-    List<String> findImagesByProductId(@Param("productId") Long productId);
+    List<String> findImagesByProductId(@Param("productId") Byte productId);
     @Query("SELECT p FROM Product p WHERE p.name LIKE %:name%")
     List<Product> findByNameContaining(@Param("name") String name);
     @Query("SELECT p FROM Product p WHERE p.name LIKE %:name%")
